Removes provider block so module can be used directly as a root config
When the provider block is present, if the module is used directly
as a root config where the user is otherwise providing a provider
config (e.g. using terragrunt generate blocks), terraform fails
with an error similar to:

```
Error: Duplicate provider configuration

  on main.tf line 3:
   3: provider aws {

A default (non-aliased) provider configuration for "aws" was already given at
main.tf:1,1-13. If multiple configurations are required, set the "alias"
```