Restricts payer configuration so only Requester is valid value
The AWS API for this action is a bit odd. The S3 default is BucketOwner,
and there is no "Delete" action for the configuration. When terraform
destroys this resource, it restores the S3 default configuration. If the
user has specified the S3 default of BucketOwner, when terraform goes
to destroy the resource, it sometimes generates a 404 saying the config
cannot be found. But since the S3 default is BucketOwner, we can simply
prevent the user from specifying that value in the first place, so we
never run into that edge case.